doc: Minor grammatical changes and flow improvements
Grammar:

Line 49: There shouldn't be a period at the end of a phrase.

Lines 56, 57, 116, 137, and 177: Adding necessary commas

Lines 103 and 136: Run-on sentence issues fixed. 

Line 176: Fixed punctuation and added necessary conjunction

Line 178: Singular noun when it should be plural

Flow:

Line 49: Adding "for" makes it more natural. 

Line 54: Though it's not grammatically incorrect to end a sentence with a preposition in an informal document such as this, the word "followed" is much easier to understand anyway, especially for those who don't have English as their native language.