[addrman] Fix new table bucketing during unserialization
An addrman entry can appear in up to 8 new table buckets. We store this
entry->bucket indexing during shutdown so that on restart we can restore
the entries to their correct buckets.

Commit 8db7aa16bf4ba54fe08a4e6ccd7e3151bd0d773e broke the
deserialization code so that each entry could only be put in up to one
new bucket. Fix that.