Halve Block Size (Block Time is /4, so Capacity is still 2x Bitcoin's).
Such limit is currently well beyond what is actually needed, and will probably be for the foreseeable future, though it means that we have lot of room as Riecoin gets adopted.
The limit can always be bumped later if it is technically needed for some useful Layer 2 implementations.

Up to 23.04, Block Size was instead doubled, providing an even more overkill 8x Capacity, while allowing spam attacks to be potentially much more harmful.
The change will be made only effective some time after 24.04 Release to give time for major actors to upgrade on time.

Signed-off-by: Pttn <28868425+Pttn@users.noreply.github.com>