refactor: remove unused `CKey::Negate` method
This method was introduced as a pre-requirement for the v2 transport
protocol back then (see PR #14047, commit 7840d504), when it was still
BIP151. With the replacement BIP324, this is not needed anymore, and
it's also unlikely that any other proposal would need to negate private
keys at this abstraction level.
(If there is really demand, it's trivial to reintroduce the method.)