test: Remove intermittent and presumed fixed tsan race suppressions
LoadWallet was added in commit 4b8c1ff2f835c0cba309a367f96f1969d8b18b3a, which
points to a traceback with BerkeleyBatch in it. Now that BDB is removed,
this can be removed as well.

The race in DatabaseBatch was added in commit
dbe8803c081e955ffdccf2c08309c32e68556951, which does not point to a
traceback. This was likely also fixed with the BDB removal.

If not, the suppressions should be added back, mentioning that they are
intermittent and including a traceback and possibly steps to reproduce.